Change CipString length to size_t.
This was done to resolve integer conversion warnings, e.g.,
Visual Studio C4267, and size_t is a more suitable data type
once the string has been decoded.

Other string types, e.g., CipShortString, should probably be
similarly altered, however, only CipString was generating
compiler warnings in WideToCipString on x64 VisualStudio.

+++ source/src/opener_api.h
@@ -449,7 +449,7 @@ int DecodeCipLreal(CipLreal *const data,
 		CipMessageRouterRequest *const message_router_request,
 		CipMessageRouterResponse *const message_router_response);
 
-int DecodeCipString(CipString *const dst,
+size_t DecodeCipString(CipString *const dst,
 		CipMessageRouterRequest *const message_router_request,
 		CipMessageRouterResponse *const message_router_response);
